Move `get_product_list` to TessDataInterface
