+++ ramjet/photometric_database/tess_data_interface.py
@@ -2,6 +2,7 @@
 Code for a class for common interfacing with TESS data, such as downloading, sorting, and manipulating.
 """
 import pandas as pd
+from astropy.table import Table
 from astroquery.mast import Observations
 from astroquery.exceptions import TimeoutError as AstroQueryTimeoutError
 
@@ -29,3 +30,21 @@ class TessDataInterface:
             except (AstroQueryTimeoutError, ConnectionError):
                 print('Error connecting to MAST. They have occasional downtime. Trying again...')
         return tess_observations.to_pandas()
+
+    @staticmethod
+    def get_product_list(observations: pd.DataFrame) -> pd.DataFrame:
+        """
+        A wrapper for MAST's `get_product_list`, allowing the use of Pandas DataFrames instead of AstroPy Tables.
+        Retries on error when communicating with the MAST server.
+
+        :param observations: The data frame of observations to get. Will be converted from DataFrame to Table for query.
+        :return: The data frame of the product list. Will be converted from Table to DataFrame for use.
+        """
+        data_products = None
+        while data_products is None:
+            try:
+                # noinspection SpellCheckingInspection
+                data_products = Observations.get_product_list(Table.from_pandas(observations))
+            except (AstroQueryTimeoutError, ConnectionError):
+                print('Error connecting to MAST. They have occasional downtime. Trying again...')
+        return data_products.to_pandas()
