+++ ramjet/photometric_database/tess_transit_lightcurve_label_per_time_step_database.py
@@ -62,24 +62,6 @@ class TessTransitLightcurveLabelPerTimeStepDatabase(LightcurveLabelPerTimeStepDa
         """
         return example_path in self.meta_data_frame['lightcurve_path'].values
 
-    @staticmethod
-    def get_product_list(observations: pd.DataFrame) -> pd.DataFrame:
-        """
-        A wrapper for MAST's `get_product_list`, allowing the use of Pandas DataFrames instead of AstroPy Tables.
-        Retries on error when communicating with the MAST server.
-
-        :param observations: The data frame of observations to get. Will be converted from DataFrame to Table for query.
-        :return: The data frame of the product list. Will be converted from Table to DataFrame for use.
-        """
-        data_products = None
-        while data_products is None:
-            try:
-                # noinspection SpellCheckingInspection
-                data_products = Observations.get_product_list(Table.from_pandas(observations))
-            except (AstroQueryTimeoutError, ConnectionError):
-                print('Error connecting to MAST. They have occasional downtime. Trying again...')
-        return data_products.to_pandas()
-
     def download_products(self, data_products: pd.DataFrame) -> pd.DataFrame:
         """
          A wrapper for MAST's `download_products`, allowing the use of Pandas DataFrames instead of AstroPy Tables.
